requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy[asyncio]
psycopg2-binary
aiohttp
numpy
pandas
scikit-learn
scipy
xgboost
torch
torch-geometric
pymc
arviz
pytensor
transformers
sentence-transformers
optuna
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
tenacity==8.2.3
playwright==1.40.0
aiosqlite
alembic
asyncpg
python-dotenv
python-multipart
aiohttp
aiosqlite
alembic
asyncpg
beautifulsoup4
fastapi
httpx
lxml
numpy
pandas
pydantic
python-dotenv
python-multipart
scikit-learn
scipy
sqlalchemy[asyncio]
tenacity
uvicorn[standard]
xgboost
aiohttp
aiosqlite
alembic
asyncpg
beautifulsoup4
fastapi
httpx
lxml
numpy
pandas
pydantic
python-dotenv
python-multipart
scikit-learn
scipy
sqlalchemy
tenacity
uvicorn
xgboost
aiohttp
aiosqlite
fastapi
httpx
python-dotenv
uvicorn
pytest
pytest-asyncio
